Still fixing font awesome icons
Issue is the icon goes the same colour as the background, so having a white background is making the icons hard to see. could darken the white but ten is it still accessible?

I've made the colours a light grey instead to make the image stand out, don't know if that's going to make the site look to dark though